requirements.txt
paho-mqtt
python-osc